refactor: separate polarized capacitor detection
